refactor: move spec to top level of project
