+++ src/bdf/__init__.py
@@ -418,7 +418,7 @@ def _csv_header_has_bdf_required(path: Path) -> bool:
         return False
     cols_l = {c.strip().lower() for c in header.split(",")}
     # import lazily to avoid cycles
-    from .normalize import spec
+    from . import spec
     for q, s in spec.COLUMNS.items():
         if not s.get("required") or bool(s.get("deprecated")):
             continue
@@ -510,7 +510,7 @@ def validate(
                         head = "".join([f.readline() for _ in range(2)]).lower()
                     header_line = head.splitlines()[0] if head else ""
                     cols_l = {c.strip().lower() for c in header_line.split(",")}
-                    from .normalize import spec
+                    from . import spec
                     for q, s in spec.COLUMNS.items():
                         if not s.get("required") or bool(s.get("deprecated")):
                             continue
