+++ src/bdf/spec.py
