+++ src/bdf/validate.py
@@ -6,10 +6,11 @@ from typing import Any, Dict, List
 
 import pandas as pd
 
-from .normalize import OPTIONAL, REQUIRED, spec
+from .normalize import OPTIONAL, REQUIRED
 from .ontology_labels import load_alias_index
 from .repair import _compute_eps_from_diffs  # reuse your epsilon heuristic
 from .units import parse_from_header
+from . import spec
 
 __all__ = ["BDFValidationError", "validate_df"]
